Remove vehicles and stations when we lost their connection info.
Due to running out of time saving, or possibly also because of a bug,
we may not always save all our connections.
To restore the connections we would have to figure out the path meaning
pathfinding. For now I don't want to bother with that.

Instead we will remove all vehicles that don't belong to a connection,
and then delete all stations without vehicles that don't belong to a
connection.
That way these vehicles won't turn up waiting in depots without ever
being deleted.
Removing stations is especially important when infrastructure maintenance
is enabled. In that case we also really should remove rail and road tracks
but I leave that for another time.