Change/Fix: TownConnectionNode::GetTownTiles
Fix a crash due to excessive CPU usage because we didn't limit town spread
to tiles belonging to town and we didn't set a maximum, causing the size
to become too big to handle in a reasonable time.
We change this by
- checking it is within the town influence
- checking it is within a spread limit we set (currently 150 tiles)

We also improve this function by adding cached values that will be
invalidated after a year. This way we don't have to recompute these values
all the time. Since the actual tiles within this area will still be checked
and the town size is not likely to change dramatically within a year, this
should not have a negative effect on our finding suitable spots for a
station.